Algorithm to solve Differential equations

Euler Method

UsefulFunctions.euler — Function

Solves (system of) Ordinary differential equations using the Euler method. function euler(F,x0,t1,t2,n,p)

Args: F (function): Function that returns the derivatives of the state vector. x0 (array): Initial state vector. t1 (float): Initial time. t2 (float): Final time. n (int): Number of steps. p (array): Parameters.

4th order Runge-Kutta Method

UsefulFunctions.rk4 — Function

Solves (system of) Ordinary differential equations using the Runge-Kutta method. function rk4(F,x0,t1,t2,n,p)

Args: F (function): Function that returns the derivatives of the state vector. x0 (array): Initial state vector. t1 (float): Initial time. t2 (float): Final time. n (int): Number of steps. p (array): Parameters.