Filters

Collection of different filters to smoothen data. For examples,

UsefulFunctions.savitzky_golay_filter_53 — Function

savitzky_golay_filter_53(data::Vector{Float64})

Applies a 5-point, 3rd-order Savitzky-Golay filter to the input data vector.

UsefulFunctions.moving_average_filter — Function

moving_average_filter(data::Vector{Float64}, window_size::Int)

Applies a moving average filter to the input data vector with the specified window size.

There are also a few numerical differentiation routine using filters,

UsefulFunctions._derivative — Function

_derivative(f, x; h=1e-2)

Computes the first derivative of function f at point x using a five-point stencil method.

UsefulFunctions._second_derivative — Function

_second_derivative(f, x; dx=1e-2)

Computes the second derivative of function f at point x using central difference method.

UsefulFunctions._sg_filter_second_derivative — Function

_sg_filter_second_derivative(f, x; dx=1e-2)

Computes the second derivative of function f at point x using a Savitzky-Golay filter. See wikipedia for details.

UsefulFunctions._first_derivative_smooth — Function

_first_derivative_smooth(f, x; dx=1e-2)

Computes a smoothed first derivative of function f at point x using a Savitzky-Golay filter. See wikipedia for details.