Integration Routines

In most cases I use pacakges like QuadGK or Cubature and HCubature for numerical integration in Julia. However, sometimes I need to use simple integration routines for quick calculations. Here are such routines

UsefulFunctions.trapizoid — Function

trapizoid(f::Function, a::Float64, b::Float64, n::Int64)

Returns the integral of f from a to b using the trapezoidal rule with n steps.

UsefulFunctions.simpson — Function

simpson(f::Function, a::Float64, b::Float64, n::Int64)

Returns the integral of f from a to b using Simpson's rule with n steps.

Principal Value Integration

UsefulFunctions.PVintegral — Function

PVintegral(f::Function, a, b, c, integrate::Function)

Calculates the Cauchy principal value integral ∫ₐᵇf(x)/(x-c)dx First it turns into a symmetric integral from -1 to 1, then substracts the singulatity. It uses the function integrate to calculate all the finite integrals.