Miscellaneous Functions

Number Distribution

UsefulFunctions.numberF — Function

numberF(temp::Float64, μ::Float64, Energy::Float64)

Returns the number density at a given Energy, chemical potential μ and temp for Fermionic Species.

UsefulFunctions.numberB — Function

numberB(temp::Float64, μ::Float64, Energy::Float64)

Returns the number density at a given Energy, chemical potential μ and temp for Bosonic Species.

Convinience constructions

UsefulFunctions.DiracDelta — Function

DiracDelta(input::Float64,δ::Float64 = 1e-3)

Dirac Delta Function that can be used in a numerical integration.

UsefulFunctions.PrincipalValue — Function

PrincipalValue(x::Float64,ϵ::Float64 = 1e-3)

Useful function in Principal Value Integration.

Unorganized

UsefulFunctions.fzero — Function

fzero(func::Function, start::Float64, finish::Float64, iteration::Int64=20)

Returns the value of the argument between start and finish where the given function func becomes zero using midpoint method.

UsefulFunctions.threaded_loop — Function

threaded_loop(iter, func_eval)

This function returns loops over the iter and evaluate the function func_eval for each value in multiple threads and returns an array of result.

UsefulFunctions._terms — Function

terms(f::Function,Λ::Float64,n::Int64)

Returns the terms for the separable representation of the function f given by f(p,p') = ∑ᵢ gᵢ(p)gᵢ(p') where gᵢ are the terms returned by this function.

Arguments

f::Function: The function to be represented.

Λ::Float64: Momentum cutoff.

n::Int64: Number of terms in approximation to be returned.