Root Finding Algorithms

Bisection Method

UsefulFunctions.bisection — Function

bisection(func::Function,start::Number,finish::Number,iteration::Integer=20)

Finds the zero of a function func inside a given interval (start,finish). Might fail if there are multiple zeros or no zeros inside the interval.

Newton-Raphson

UsefulFunctions.NewtonRaphson — Function

NewtonRaphson(f::Function,initial_guess;tol=1e-3,ϵ=1e-3,maxiter=20)

Uses Newton Raphson Method to find zero of a given function f starting from an initial_guess.

\[x_{n+1} = x_n - \frac{f(x_n)}{f'(x_n)}\]

where discreet derivatives are used when functional form of the derivative are not given as input.