Root Finding Algorithms

Bisection Method

UsefulFunctions.bisection — Function

bisection(func::Function,start::Number,finish::Number,iteration::Integer=20)

Finds the zero of a function func inside a given interval (start,finish). Might fail if there are multiple zeros or no zeros inside the interval.

Newton's Method

UsefulFunctions.NewtonMethod — Function

NewtonMethod(func::Function, Jacobian::Function,initial_guess::Array,iteration=100)

uses the Newton method toi find zeros in more than one dimensions. $x_{n+1} = x_n - J^{-1}(x_n)f(x_n)$

UsefulFunctions.NewtonRaphson — Function

NewtonRaphson(f::Function,initial_guess;tol=1e-3,ϵ=1e-3,maxiter=20)

Uses Newton Raphson Method to find zero of a given function f starting from an initial_guess.

\[x_{n+1} = x_n - \frac{f(x_n)}{f'(x_n)}\]

where discreet derivatives are used when functional form of the derivative are not given as input.

Brent's Method

UsefulFunctions.brent — Function

brent(f::Function, x0::Number, x1::Number, args::Tuple=(); xtol::AbstractFloat=1e-4, ytol=1e-6, maxiter::Integer=20)

The function uses Brent's Method to find zero inside the given interval (x0,x1). Might fail if there are multiple zeros inside the interval.

Broyden's Method

UsefulFunctions.broyden — Function

broyden(fun, jaco, x, iter=50, ftol=1e-7, verbose=false)

Uses Broyden's method to find zero of a multidimensional function.